Generated by Spark: Now we need a section to genarate a bracket style fighting tournament for the championship match. This will be single elimination. It will be the top 32 fighters
And it will also be auto placements and best matched with stats or we should have the option to randomize the placements for the best odds.

Figure out what is traditionally used and go with that and make sure its fair

Also on the registration of the boxing add in a check box to see what their timezone is.
Use NA/EU/AU for the options