Generated by Spark: So the betting needs to be changed to where we are the bookings
So we would collect state IDs and place their bets for them.

The way the odds are are okay

But I need the fights to have a way to set the maximum manually

Then we would need a way to show how much we pay the boxers.

So the person betting would get the biggest cut

The booker would bet a cut

LSBA gets a cut

the boxer gets a cut

and it can be very small % but we need to also have a way to change and manage it if needed

I think LSBA should get a cut of every bet and so should the sponsor, at minimum since the betting money would be managed byt the casino called 1068